requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
idna==3.3
newsapi-python==0.2.6
PyMySQL==1.0.2
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
